school_management_backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
python-dotenv
